refactor(quality): raise fallow code-health grade D→C
Lifts the repo's tracked code-health tier one grade — fallow maintainability
goes 51.6/D → 60.8/C (above the committed snapshot's stale 59/C). Three prongs,
each a genuine improvement or a verified false-positive correction:

1. De-duplicate the sandbox-sim fetch-shim adapters. browser.ts and node.ts were
   byte-identical bar comments (a 100% clone). Extract the shared core into
   adapters/fetch-shim.ts, decomposing the cognitive-33 `toSimRequest` into small
   helpers (normalizeUrl/Method/Headers, readBody, drainStream), and reduce the
   two adapters to thin re-export entry points — preserving the documented
   "specialise here for a divergent body reader" seam. Kills the clone, two
   complexity hotspots, and the duplicated large unit. Behaviour-preserving.

2. Break the docs/sandbox/contracts import cycle. dispatch.ts was both the frozen
   contract (types + NODE_ONLY_SURFACES) and a barrel re-exporting the impls,
   which imported the contract back. Move the primitives to a dependency-free leaf
   (surface.ts); dispatch.ts re-exports it, so the frozen public surface is
   unchanged, and the impls now import the leaf. The lazy-init workaround the cycle
   forced in scan-surface.ts becomes a plain module const.

3. Add an accurate .fallowrc.jsonc. The repo ran fallow with zero config, so its
   own `pnpm metrics`/README badge over-reported false positives: live files (the
   custom run-all.mjs test suites, the esbuild worker/MCP/stitch-browser entry
   points) flagged as dead, and deps used at runtime / via deep relative paths
   flagged as unused. Every entry is evidence-backed (a test the runner executes
   or an esbuild entry a build script bundles); the two ignored deps are ones
   fallow itself reports as used elsewhere. unused_files 24→0, circular 2→0, with
   no real dead code masked.

Refresh the README health badge from the now-accurate score (59→61, both C).

No public API, dependency, or lockfile change. Verified: sandbox-sim + eval-harness
+ apps/docs typecheck, sandbox-sim/eval-harness tests, docs/sandbox 11/11 smoke,
prettier --check, eslint, README drift guard.
