+++ README.md
@@ -23,7 +23,7 @@
 
 <p align="center">
   <a href="https://stand-with-ukraine.pp.ua"><img alt="StandWithUkraine" src="https://raw.githubusercontent.com/vshymanskyy/StandWithUkraine/main/badges/StandWithUkraine.svg" /></a>
-  <img alt="code health: 59 (C)" src="https://img.shields.io/badge/code_health-59_%28C%29-yellow" />
+  <img alt="code health: 61 (C)" src="https://img.shields.io/badge/code_health-61_%28C%29-yellow" />
   <img alt="coverage: 90% lines · 78% branches" src="https://img.shields.io/badge/coverage-90%25_lines_%C2%B7_78%25_branches-green" />
   <a href="LICENSE"><img alt="license: Apache-2.0" src="https://img.shields.io/badge/license-Apache--2.0-blue" /></a>
 </p>
